% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summarize_target.R
\name{summarize_target}
\alias{summarize_target}
\title{Summarize a Target Numeric Variable in a Data Frame by Specified Grouping Variables}
\usage{
summarize_target(data, group_vars, target_var)
}
\arguments{
\item{data}{A data frame containing the data to be summarized.
Named 'data' to indicate this part of function expects a data frame as input.}

\item{group_vars}{A character vector specifying the names of the columns to group by.
Named 'group_vars' to indicate these variables determine the grouping structure}

\item{target_var}{A single character string specifying the name of the numeric column to summarize.
Named 'target_var' to indicate this is the target variable need to be summarized.}
}
\value{
A data frame containing the grouped summary statistics, including mean, minimum, maximum, and
standard deviation of the target variable for each combination of grouping variables.
}
\description{
This function calculates summary statistics for a specific numeric column in a data frame. It computes the
mean, minimum, maximum, and standard deviation for each group defined by the specified grouping variables.
}
\examples{
# Use the steam_games data here to compute summary statistics for the
# discount_price variable group by types.
summarize_target(datateachr::steam_games, "types", "discount_price")

# The function will stop and display an error message
# indicating that the specified grouping variable does not exist in the data frame.
try(summarize_target(datateachr::steam_games, "color", "discount_price"))
}
